+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/Ipv4Validator.kt
@@ -1,5 +1,7 @@
 package com.atlassian.performance.tools.awsinfrastructure
 
-internal class Ipv4Validator : (String) -> Boolean {
-    override fun invoke(ip: String) = ip.matches(Regex("[0-9]+\\.[0-9]+\\.[0-9]+\\.[0-9]+"))
+import java.util.function.Predicate
+
+internal class Ipv4Validator : Predicate<String> {
+    override fun test(ip: String) = ip.matches(Regex("[0-9]+\\.[0-9]+\\.[0-9]+\\.[0-9]+"))
 }