+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/network/access/ForCidrAccessRequester.kt
@@ -1,10 +1,12 @@
 package com.atlassian.performance.tools.awsinfrastructure.api.network.access
 
+import java.util.function.Supplier
+
 class ForCidrAccessRequester(
-    private val cidrProvider: () -> String
+    private val cidrProvider: Supplier<String>
 ) : AccessRequester {
     override fun requestAccess(
         accessProvider: AccessProvider
     ) = accessProvider
-        .provideAccess(cidrProvider())
+        .provideAccess(cidrProvider.get())
 }