+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/network/access/ForSecurityGroupAccessRequester.kt
@@ -1,12 +1,13 @@
 package com.atlassian.performance.tools.awsinfrastructure.api.network.access
 
 import com.amazonaws.services.ec2.model.SecurityGroup
+import java.util.function.Supplier
 
 class ForSecurityGroupAccessRequester(
-    private val securityGroupProvider: () -> SecurityGroup
+    private val securityGroupProvider: Supplier<SecurityGroup>
 ) : AccessRequester {
     override fun requestAccess(
         accessProvider: AccessProvider
     ) = accessProvider
-        .provideAccess(securityGroupProvider())
+        .provideAccess(securityGroupProvider.get())
 }