+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/virtualusers/ProvisionedVirtualUsers.kt
@@ -11,30 +11,15 @@ class ProvisionedVirtualUsers<out T : VirtualUsers> private constructor(
     val resource: Resource,
     val accessRequester: AccessRequester
 ) {
-    object Defaults {
-        val resource: Resource = UnallocatedResource()
-        val accessRequester: AccessRequester = NoAccessRequester()
-    }
-
-    @Deprecated("Use ProvisionedVirtualUsers.Builder instead.")
-    constructor(
-        virtualUsers: T,
-        resource: Resource
-    ) : this(
-        virtualUsers = virtualUsers,
-        resource = resource,
-        accessRequester = Defaults.accessRequester
-    )
-
-    override fun toString(): String {
+     override fun toString(): String {
         return "ProvisionedVirtualUsers(virtualUsers=$virtualUsers, resource=$resource)"
     }
 
     class Builder<out T : VirtualUsers>(
         private val virtualUsers: T
     ) {
-        private var resource: Resource = Defaults.resource
-        private var accessRequester: AccessRequester = Defaults.accessRequester
+        private var resource: Resource = UnallocatedResource()
+        private var accessRequester: AccessRequester = NoAccessRequester()
 
         fun resource(resource: Resource) = apply { this.resource = resource }
         fun accessRequester(accessRequester: AccessRequester) = apply { this.accessRequester = accessRequester }